Split the tree concurrency test codes
